+++ packages/orm/angel_orm_mysql/test/migrations/leg.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE legs (
+CREATE TABLE IF NOT EXISTS legs (
     id serial PRIMARY KEY,
     name varchar(255) NOT NULL,
     created_at timestamp,