+++ packages/orm/angel_orm_mysql/test/migrations/user.sql
@@ -1,4 +1,4 @@
-CREATE TEMPORARY TABLE users (
+CREATE TABLE IF NOT EXISTS users (
   id serial PRIMARY KEY,
   username varchar(255),
   password varchar(255),