+++ packages/orm/angel_orm_mysql/test/migrations/weird_join.sql
@@ -1,15 +1,15 @@
-CREATE TEMPORARY TABLE weird_joins (
+CREATE TABLE IF NOT EXISTS weird_joins (
   id serial,
   join_name varchar(255),
   PRIMARY KEY(id)
 );
 
-CREATE TEMPORARY TABLE foos (
+CREATE TABLE IF NOT EXISTS foos (
   bar varchar(255) not null UNIQUE,
   PRIMARY KEY(bar)
 );
 
-CREATE TEMPORARY TABLE foo_pivots (
+CREATE TABLE IF NOT EXISTS foo_pivots (
   weird_join_id int,
   foo_bar varchar(255)
 );