+++ packages/orm/angel_orm_mysql/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ## 6.0.0-beta.3
 
-* Updated transaction for `MariaDbExecutor`
-* Updated transaction for `MySqlExecutor`
+* Fixed transaction for `MariaDbExecutor`
+* Fixed transaction for `MySqlExecutor`
+* Fixed error for non `id` primary key
+* Changed test cases to use tables instead of temporary tables to overcome limitations
 
 ## 6.0.0-beta.2